requirements.txt
greenlet==1.1.2
psycopg2==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.39
